+++ Sora/Signaling.swift
@@ -408,6 +408,9 @@ public struct SignalingOffer {
     /// scalability mode
     public let scalabilityMode: String?
 
+    /// ネットワーク優先度 (DiffServ Code Point)
+    public let networkPriority: RTCPriority?
+
     /// RTP エンコーディングに関するパラメーター
     public var rtpEncodingParameters: RTCRtpEncodingParameters {
       let params = RTCRtpEncodingParameters()
@@ -426,6 +429,9 @@ public struct SignalingOffer {
         params.scaleResolutionDownTo?.maxHeight = value.maxHeight
       }
       params.scalabilityMode = scalabilityMode
+      if let value = networkPriority {
+        params.networkPriority = value
+      }
       return params
     }
   }
@@ -1075,6 +1081,7 @@ extension SignalingOffer.Encoding: Codable {
     case scaleResolutionDownBy
     case scaleResolutionDownTo
     case scalabilityMode
+    case networkPriority
   }
 
   public init(from decoder: Decoder) throws {
@@ -1099,6 +1106,24 @@ extension SignalingOffer.Encoding: Codable {
     scalabilityMode = try container.decodeIfPresent(
       String.self,
       forKey: .scalabilityMode)
+    if let rawNetworkPriority = try container.decodeIfPresent(
+      String.self, forKey: .networkPriority)
+    {
+      switch rawNetworkPriority {
+      case "very-low":
+        networkPriority = .veryLow
+      case "low":
+        networkPriority = .low
+      case "medium":
+        networkPriority = .medium
+      case "high":
+        networkPriority = .high
+      default:
+        networkPriority = nil
+      }
+    } else {
+      networkPriority = nil
+    }
   }
 
   public func encode(to encoder: Encoder) throws {